int64 for wheel_vtarget (for applying to ros controller)

+++ src/twist_to_motors.cpp
@@ -48,9 +48,9 @@ TwistToMotors::TwistToMotors()
 
     cmd_vel_sub = n.subscribe("cmd_vel", 10, &TwistToMotors::twistCallback, this);
 
-    pub_lmotor = n.advertise<std_msgs::Float32>("lwheel_vtarget", 50);
+    pub_lmotor = n.advertise<std_msgs::Float64>("lwheel_vtarget", 50);
 
-    pub_rmotor = n.advertise<std_msgs::Float32>("rwheel_vtarget", 50);
+    pub_rmotor = n.advertise<std_msgs::Float64>("rwheel_vtarget", 50);
 }
 
 void TwistToMotors::init_variables()
